Compatibility : WooCommerce - Gravity Forms Product Add-Ons
This plugin became incompatible with WGF in version 2.6.0 with the addition of

var form_id = jQuery( this ).parents( 'form' ).attr( 'id' ).split( '_' ).pop().trim();

added after the:

jQuery( 'input.itsg_ajax_upload_browse' ).each(function () {

This was done to remove similar looking statements down inside callbacks built
in the each. It works fine in regular Gravity Forms, but in WooCommerce Gravity
Forms, somehow the form id would either not be set up yet, or removed before,
when this statement runs. So when this statement ran it triggers a fatal
javascript error from calling split() on an undefined value and the entire form
would not render.

However, this statement is not necessary. The form_id, along with other form
level variables, are fetched before the iterator and is already available. So
this change removes this statement allowing the callbacks to use the form_id
set outside the iterator.